Feat: 덱 조회 API 구현
및 관련 Entity, Dto 작성

+++ backend/src/main/java/com/ssafish/domain/deck/Card_deckRepository.java
@@ -2,9 +2,17 @@ package com.ssafish.domain.deck;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Card_deckRepository extends JpaRepository<Card_decks, Integer> {
+
+public interface Card_deckRepository extends JpaRepository<Card_decks, Long> {
+
+    //deckId 와 매핑된 모든 cardId를 반환한다.
+  //  List<Long> findAllCardIdByDeckId(long deckId);
 
 
 
 
+    //List<Card> findCardDeckList(long deckId);
 }
+
+
+
